+++ src/c/so3.h
@@ -6,5 +6,6 @@
 #include "so3_error.h"
 #include "so3_sampling.h"
 #include "so3_core.h"
+#include "so3_conv.h"
 
 #endif // SO3_H
