+++ src/c/so3_sampling.h
@@ -25,11 +25,12 @@ double so3_sampling_g2gamma(int g, const so3_parameters_t* parameters);
 
 // Note, if this is compiled using C99-standard then the "extern" belongs in the
 // .c file instead.
-extern inline int so3_sampling_flmn_size(const so3_parameters_t* parameters);
-extern inline void so3_sampling_elmn2ind(int* ind, int el, int m, int n, const so3_parameters_t* parameters);
-extern inline void so3_sampling_ind2elmn(int* el, int* m, int* n, int ind, const so3_parameters_t* parameters);
-extern inline void so3_sampling_elmn2ind_real(int* ind, int el, int m, int n, const so3_parameters_t* parameters);
-extern inline void so3_sampling_ind2elmn_real(int* el, int* m, int* n, int ind, const so3_parameters_t* parameters);
+int so3_sampling_flmn_size(const so3_parameters_t* parameters);
+void so3_sampling_elmn2ind(int* ind, int el, int m, int n, const so3_parameters_t* parameters);
+void so3_sampling_ind2elmn(int* el, int* m, int* n, int ind, const so3_parameters_t* parameters);
+void so3_sampling_elmn2ind_real(int* ind, int el, int m, int n, const so3_parameters_t* parameters);
+void so3_sampling_ind2elmn_real(int* el, int* m, int* n, int ind, const so3_parameters_t* parameters);
+void so3_sampling_n_loop_values(int *n_start, int *n_end, int *n_step, const so3_parameters_t *parameters);
 
 #ifdef __cplusplus
 }
